+++ src/Opulence/Framework/Console/Testing/PhpUnit/Assertions/ResponseAssertions.php
@@ -11,12 +11,12 @@ namespace Opulence\Framework\Console\Testing\PhpUnit\Assertions;
 use LogicException;
 use Opulence\Console\Responses\StreamResponse;
 use Opulence\Console\StatusCodes;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 
 /**
  * Defines the console response assertions
  */
-class ResponseAssertions extends PHPUnit_Framework_TestCase
+class ResponseAssertions extends TestCase
 {
     /** @var StreamResponse The response stream */
     protected $response = null;