+++ src/Opulence/Framework/Http/Testing/PhpUnit/Assertions/ViewAssertions.php
@@ -10,12 +10,12 @@ namespace Opulence\Framework\Http\Testing\PhpUnit\Assertions;
 
 use LogicException;
 use Opulence\Routing\Controller;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 
 /**
  * Defines the HTTP view assertions
  */
-class ViewAssertions extends PHPUnit_Framework_TestCase
+class ViewAssertions extends TestCase
 {
     /** @var Controller|mixed The matched controller */
     protected $controller = null;