+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+<h2>v1.0.3 (2017-01-?)</h2>
+
+<h3>Framework</h3>
+* Updated the various `Assertions` classes to extend the namespaced `PHPUnit\Framework\TestCase` class rather than the soon-to-be-outdated `PHPUnit_Framework_TestCase`
+
 <h2>v1.0.2 (2016-12-16)</h2>
 
 <h3>Console</h3>